# apps/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Owostack API",
    "description": "Unified billing, entitlements, and usage metering API. Owostack sits between your app and payment providers (Paystack, Stripe, Dodo, Polar) to handle subscriptions, feature gating, metered usage, credit systems, and invoicing.",
    "version": "1.0.0",
    "contact": {
      "name": "Owostack",
      "url": "https://owostack.com"
    }
  },
  "servers": [
    {
      "url": "https://api.owostack.com",
      "description": "Production"
    },
    {
      "url": "https://sandbox.owostack.com",
      "description": "Sandbox"
    }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/v1/subscriptions/{id}/checkout": {
      "post": {
        "operationId": "generateSubscriptionCheckout",
        "tags": ["Subscriptions"],
        "summary": "Generate checkout for a pending subscription",
        "description": "Generates a payment checkout URL for an existing pending subscription. Useful for collecting payment details before activating.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Subscription ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "callbackUrl": { "type": "string", "format": "uri", "description": "Redirect URL after successful payment setup" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "checkoutUrl": { "type": "string", "format": "uri" },
                    "reference": { "type": "string" },
                    "accessCode": { "type": "string" },
                    "activatedDirectly": { "type": "boolean", "description": "True if the plan was free and was activated immediately without checkout." }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/subscriptions/{id}/activate": {
      "post": {
        "operationId": "activatePendingSubscription",
        "tags": ["Subscriptions"],
        "summary": "Force activate a pending subscription",
        "description": "Manually marks a pending subscription as active without a checkout. Bypasses payment requirements. Typically used for offline sales or full 100% discounts.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Subscription ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription activated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/v1/plans": {
      "get": {
        "operationId": "listPlans",
        "tags": ["Plans"],
        "summary": "List all plans",
        "description": "List all plans for the organization with optional filters.",
        "parameters": [
          { "name": "group", "in": "query", "schema": { "type": "string" }, "description": "Filter by plan group" },
          { "name": "interval", "in": "query", "schema": { "type": "string" }, "description": "Filter by billing interval" },
          { "name": "currency", "in": "query", "schema": { "type": "string" }, "description": "Filter by currency (case-insensitive)" },
          { "name": "includeInactive", "in": "query", "schema": { "type": "boolean" }, "description": "Include inactive plans" }
        ],
        "responses": {
          "200": {
            "description": "Plan list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "plans": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Plan" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/plans/{slug}": {
      "get": {
        "operationId": "getPlan",
        "tags": ["Plans"],
        "summary": "Get a plan by slug",
        "description": "Retrieve a single plan with all its features.",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Plan details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "plan": { "$ref": "#/components/schemas/Plan" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key in the format `owo_sk_...`"
      }
    },
    "schemas": {
...
      "Plan": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "slug": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "price": { "type": "number" },
          "currency": { "type": "string" },
          "interval": { "type": "string" },
          "type": { "type": "string", "enum": ["free", "paid"] },
          "billingType": { "type": "string" },
          "isAddon": { "type": "boolean" },
          "planGroup": { "type": "string", "nullable": true },
          "trialDays": { "type": "integer" },
          "provider": { "type": "string", "nullable": true },
          "features": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "slug": { "type": "string" },
                "name": { "type": "string" },
                "type": { "type": "string", "enum": ["metered", "boolean"] },
                "enabled": { "type": "boolean" },
                "limit": { "type": "number", "nullable": true },
                "resetInterval": { "type": "string", "nullable": true },
                "unit": { "type": "string", "nullable": true },
                "overage": { "type": "string", "enum": ["charge"] },
                "overagePrice": { "type": "number" }
              }
            }
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "enum": [true] }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": { "type": "string" }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "error": "Invalid API Key" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "error": "Invalid request body" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "error": "Not found" }
          }
        }
      }
    }
  }
}